backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["serramatutu <lucasvvop@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
sanic = "^21.3.4"
Sanic-Cors = "^1.0.0"
aioredis = "^1.3.1"
pyzmq = "^22.1.0"
PyYAML = "^5.4.1"
Cerberus = "^1.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
